Trigger 'woocommerce_order_add_product'
When creating renewal orders, we copy the product line items over
without using WC_Abstract_Order::add_product(), because it requires
the product to still exist, which isn't necessarily the case for
line items on renewal orders. However, this means we were not
triggering 'woocommerce_order_add_product'. This patch now triggers
that hook when the related product still exists.