Manage cancellation date in WC_Subscription
Instead of managing it in WCS_Admin_Reports (even though it's only
used for reports at the moment, it makes sense to manage it alongside
other subscription dates)..

Adds handling for the cancellation date to WC_Subscription's date and
status API methods. The end date for a subscription must now occur at
the same time or later than the cancellation date, which is enforced
by the update_dates() method.

Also display the cancelled date above the end date on the Edit
Subscription administration screen.

Fixes #1611