Fix current subscription count for monthly group
The prepare_chart_data() method aggregates all data, but the current
subscription line is an average of daily subscriptions that haven't
ended yet, so aggregating that when the report is based on monthly
grouping would lead to (very) incorrect numbers (e.g. ~30x the
correct number of current subscriptions).

This fixes that by introducing a new prepare_chart_data_daily_average()
method which will fallback to prepare_chart_data() when the chart
is grouped by day, or calculate it's own data if grouped by months
that averages the data, rather than aggregating it.