Adds filter to the list of related orders to a post
No issue.

Makes it easier to manipulate the list of related orders on a subscription / order. The alternative is to unhook the `output_rows` method from the `woocommerce_subscriptions_related_orders_meta_box_rows`, but that would mean the method would largely be duplicated anyway.

The use case is that if there are other posts / orders that need to go there, or we need to not show resubscribe orders for whatever reason.