Cache package shipping methods against the applied coupons and contents cost
Cached rates are stored against a key generated by hashing the package contents.
However, this throws up some errors where a shipping method requires a minimum order total or coupon to be available.
For instance, if free shipping requires the customer to spend $50 and the cart contains a subscription with a $50+ signup fee and recurring amount < $50.
In this case the recurring cart package contents match the initial cart contents, however the recurring amount doesn't meet the shipping method requirement but the cache would treat these two packages as the same, returning incorrect rates.