Fall back to recurring cart renewal date when there is no switching first payment date
When switching prorated totals has set the first payment date to 0, it is intended that
the subscription will start from today. However, when processing switches through the
checkout, 0 first payment dates are considered as keeping the exisiting subscription
payment dates. This patch sets the first payment date to the renewal cart's next
payment date so that it is used during the comparisons.