Only validate initial order carts where the order is a parent
Currently the initial cart setup will only check if the order doesn't contain renewal and resubscribes.
So orders which just contain simple products will be validated by maybe_setup_cart(). The cart won't actually be set up
however because it will fall over the next check. This commit just moves the check on whether an order is a parent higher.