Add trigger_gateway_renewal_payment_hook()
And move most of the requirements checks before triggering the gateway
hook to this method and run them on the renewal order instead of the
subscription (e.g. the cost of the renewal order, the gateway etc.)
to make sure the renewal orders values are honoured, regardless of
the subscription (meaning you could now have a subscription set to
charge $0 / month, but if the renewal order's total is > $0, the
payment will still be processed).