Fix PayPal renewals when ALTERNATE_WP_CRON defined
As reported in ticket https://woothemes.zendesk.com/agent/tickets/478112
when ALTERNATE_WP_CRON is defined, Subscriptions is unable to process
renewal payments for subscriptions using PayPal as the payment method.
Instead, it simply logs on the subscription the order note:
> Unable to change subscription status to "on-hold"

This is because when ALTERNATE_WP_CRON is defined, WordPress runs WP-Cron
on the 'init' hook with the default priority of 10. That means anything
hooked to 'init' with the same or higher priority will not run for those
WP-Cron requests, while they will run for normal WP-Cron requests. This is
a known issue with WordPress as reported in trac here ticket 24160:
https://core.trac.wordpress.org/ticket/24160

Subscriptions attached WC_Subscriptions_Payment_Gateways::init_paypal() to
'init' with the default priority, which was responsible for calling the
WCS_PayPal::init() method, which then called WCS_PayPal_Supports::init()
to setup the check for whether the subscription could be suspended or not
via the 'woocommerce_subscription_payment_gateway_supports' filter. Because
this was never called when ALTERNATE_WP_CRON was defined, the subscription
could not be suspended, meaning the renewal could not be processed.