Add WCS_Retry::update_status()
And use it to replace WCS_Retry_Manager::update_retry_status() to provide
a central point for updating the retry's status, rather than having an
external API function responsible for that. This is particularly important
now that the 'woocommerce_subscriptions_retry_status_updated' exists as
that hook should always be called when a retry's status is updated, not just
when the WCS_Retry_Manager API function is used to update it.