Sum parent order totals correctly
Also refactor the query to make sure that its possible to both sum
renewal order totals correctly and to count the number of subscriptions
created for each parent order (which would be more than one).

Simply adding a "SUM(wcometa.meta_value) as signup_totals" clause to the
existing query would result in adding the order total for each subscription
created in an order, because we were doing a join on the subscription IDs.
The new query joins a _count_ of the subscriptions to each order ID instead
of the IDs of each subscription so that there is only one _order_total meta
row for each order rather than one for each subscription.

Also takes the chance to use more descriptive MySQL table aliases while
changing the query.