Add switched items as _pending_switch on checkout
In preparation for removing the TRANSACTION and ROLLBACK functionality.

Instead of adding the line items, storing its meta and then rolling
back the changes. Just leave the items added but give them a different
line item type so they aren't displayed on the subscription.

When payment is received, flip them to normal line items.