Update renewal order line item cart data when item IDs are reset
When a renewal order is processed through the checkout the existing
line items are removed and new items are added from the cart.
We use the renewal order line item ids to honour pricing and quantities,
and so there exists a situation where the old line items don't exist
but we're still trying to access them. Causing errors like:
- PHP Warning:  Division by zero
- PHP Notice:  Undefined offset

This fixes that by updating the cart item data after WC has re-added
the new items.