Set the recurring cart key flag when validating recurring shipping methods
When determining if free shipping is available to a recurring cart,
we require the package to have the recurring cart key index match.

In order for the package to set the recurring cart key index it needs
the global recurring cart key flag to be set.

This commit simply sets it when we're validating the shipping methods.