Adds more flexibility to next payment date recalculation
No issue.

On a client project we replaced the action scheduler logic. Between midnight at 4 am we're charging all the subscriptions that are due to be charged on that day, regardless of what time of the day they're set to.

Because the current code has a 2 hour threshold hardcoded (magic number programming), which does not have a filter, the only viable option is to repeat the code found in WC_Subscription::update_status case "active" in the woocommerce_subscription_status_active' action hook.

That's not great because if core code changes and plugin code does not immediately notice it, it could lead to bugs.

The introduction of this filter is supposed to reduce code duplication, offer more flexibility over how subscriptions work, and adds an additional action hook in case plugin wants to do something (like generate a report) if / when a subscription was reactivated, BUT its next payment date was not recalculated.