WC2.7: Add WC_Product_Subscription_Variation_Legacy::get_parent_id()
And use it to try to consistently get the parent ID for variations when
WC < 2.7 is active. We can't use WC_Product::get_parent(), becuase in WC
prior to 2.7, the value of $this->post->post_parent for a variation would
be 0, because $this->post is actually set to the post data for the parent
variable product already.