Remove unsaved post title setting
And also deprecate wcs_get_new_order_title() because it's only
used for that purpose.

The custom post title used for renewals was being set in memory,
but not saved in the database, meaning it was a useless line of
code. The change was introduced with SHA: 29cb8c2b, which was
included in 2.0.0, so the post title's for all renewals orders
unintentionally changed between 1.n and 2.0. No one has complained
about this change, so there's no point updating all the existing
post titles and fixing this code to maintain that behaviour.