WC2.7: Make last payment date source consistent
While migrating data to use wcs_get_objects_property(), it became
clear that last payment dates were being inconsistently sourced
from two different places:

1. the post_date_gmt column on the last order; or
2. the _paid_date post meta value on the last order

This inconsistency stems from the fact that orders do not always
have a _paid_date meta value set on them. It is only set when the
payment gateway calls $order->payment_complete(), and not all gateways,
including core gateways like BACs or Cheque, call this method. :sob:

To get around this inconsistency, we often use the post_date_gmt
columns value, which is often updated when $order->payment_complete()
is called, so should be equivalent to the '_paid_date' meta value.
However, it's not sufficient to always use this date, because we
can't be sure based on it whether the order has actually been paid.
Nor can we rely on checking if the order has a paid order status,
because the order may have been paid, and transitioned to a valid
paid order status, which was then later deemed an invalid order
status.

Because of this, this patch:

1. Uses the paid date value when it exists
2. Uses the 'post_date_gmt' column or date created prop when it doesn't
2. Ensures that an order is considered paid if it has a '_paid_date'
   post meta value set, even if its order status isn't a valid paid
   order status
3. Ensures the '_paid_date' value is set when updating the last
   payment date

The last of those changes also improves compatibility with 3rd party
plugins, which may be expecting the '_paid_date' post meta to be set
correctly and using that to determine the paid date for an order.