Prevent use of WC_Order_Item_Product objects being used as cart items
When generating the PayPal RT item description args, the items return by
$order->get_items() are sent to the function get_item_description() which
can accept cart items and order items.

In WC 2.7 however, an order item isn't an array and was being treated as one.

This small change makes sure we're dealing with an array (cart item or order item pre 2.7)
before trying to access the product.