WC2.7: Fix wcs_get_objects_property multiple value
When the $single parameter is set to 'multiple' or something else
indicating we don't want just one value, the expected return value
was to match the array of values get_post_meta() would return when
its $single flag was set to false. However, WC_Data::get_meta() will
actually return an array of stdClass objects when the $single param
is set to false, not an array of values.

This patch ensures the return value aligns with the older return values.

Related to woocommerce/woocommerce#13427