WC2.7: Fix date setting
SHA: 6da1fbd6f7 fixed most things around how Subscriptions handles
dates with WC 2.7, but it missed one major issue - date props for
all subscription dates were not being set correctly, because there
was not set_{$date_type}() method in WC_Subscription, because we
want to continue to use update_dates() so that we can enforce our
date validation.

This patch updates the subscription data store to use udpate_dates()
for subscription date types. This makes sure date props are set.

An alternative fix would be to add the set_{$date_type}() methods to
WC_Subscription and let the existing read_order_data() logic handle
the date types. These should still call validate_dates() before they
set the date, and the order of the set_{$date_types}() would need to
be called in a specific order in line with the validation, which is
why for now, I've gone with the udpate_dates() method where the dates
can be passed in one array.

This will need to be rewritten once woocommerce/woocommerce#13498 is
address with woocommerce/woocommerce#13577 and other PRs, but I'm
committing it for now for posterity and to get the unit tests
working.