WC2.7: Fix requires manual renewal update
In previous versions of Subscriptions, the '_requires_manual_renewal'
post meta value was a string representation of a boolean, i.e. either
'true' or 'false' (when set). (Not the best, but better than storing and
retrieving booleans from post meta where they could go in as boolean
true and come out as string '1' or go in as boolean false and come out
as an empty string '').

To be backward compatible with that we need to:

1. store the 'requires_manual_renewal' prop in the same form so that when
   it is written to the database in post meta, the value doesn't change
2. make sure the WC_Subscription::__get() method called for the old
   $requires_manual_renewal property returns the 'true' or 'false'
   like it used it.

However, we don't need to return 'true' or 'false' strings from the new
get_requires_manual_renewal() method, because it's a new method and can
return whatever we want. Similarly with set_requires_manual_renewal(),
we can accept any params, but still store it however we want. Its better
to return/accept booleans, so we can do that and just need to make sure
the above 2 requirements are handled in the data store and setter.