WC3.0 Update last payment/order date handling
Deprecate 'last_payment' date type in favour of the more accurately
named 'last_order_date_created', which is also consistent with the
new WC 3.0 nomenclature. For calls to existing date API functions,
like WC_Subscription::get_date(), we continue to return the last order's
creation date, the same as the deceptively named get_last_payment_date()
method did previously (at least in most cases).

This patch also updates get_date_paid() and get_date_completed(), both
of which are new WC 3.0 methods on WC_Order, to return the most recent
date of that type from a related order, rather than just the date of the
most recent order. This is more accurate, because a subscription that has
a payment, just not on the most recent order, still has a paid date, its
just not the date of the most recent order. Becuase these are new functions,
we can also define their behaviour now without worrying about backward
compatibility. However, the set_date_paid() and set_date_completed()
methods _will_ act only on the most recent related order, because it makes
sense that when setting that date, we want to set it on the most recent
order (regardless of whether the date exists on that order or not yet).

To get the date paid or completed only for the last order, new date types
have also been introduced as valid keys for WC_Subscription::get_date() and
WC_Subscription::get_time():

* 'last_order_date_paid'
* 'last_order_date_completed'

Fixes #1943