Update the PUT /subscriptions endpoint to work with new CRUD (wc3.0.0)
Similarly to the create subscription endpoint changes, the update
subscription function no longer calls the parent (update_order)
function. This is because it was calculating totals and also it allows
us to have freedom in the ordering the code differently (like updating
the payment method meta, just after save() is called, but before the
subscription is potentially marked as completed.