WC3.0 Fix infinite loop on subscription init
Caused by update_dates() getting the existing dates, some of which
come from orders, which means the orders are instantiated, which
sometimes means WC_Subscriptions_Order::maybe_autocomplete_order()
is called, and if the order has a $0 total, then that also mean that
wcs_order_contains_subscription() is called, which calls the
wcs_get_subscriptions_for_order() method, which attempts to
instantiate orders relating to the order.

We could also add a method which *doesn't* init subscriptions just
to check if there are any related to the order. That would have
significant performance gains, but for now, this patch ensures we
don't bother getting dates sourced from orders, becuase that should
avoid other potential loops/issues.

Similar to #1983