Fix the status transition note when a update fails on WC 2.6
When a subscription status change fails, a order note is still
added. On WC 3.0 this note is something says active to active,
for example, however on 2.6 it would would appear as if a change
had actually taken affect when if fact it hadn't. This commit
puts the WC_Subscription_Legacy::set_status() function in line
with it's WC 3.0 counter part.