Share CRUD item ID matching in renewal & switching
SHA: d82d7e3 added code to make switching compatible with WC 3.0's
new CRUD architecture where an order item is created then saved later,
meaning when order items are created on checkout, it does not yet have
and ID, because it hasn't been saved.

This code worked well for switching, but that's not the only occasion
we need to match an order item ID. We also need to do it via renewal
to address the issues from #1795. SHA: 9378b7d81 was supposed to do
this for renewal, but failed, bad. Which led to #2026.

This patch shares the code already used for switch and located in by
WC_Subscriptions_Switcher by moving it into WCS_Cart_Renewal so that
when WCS_Cart_Switch inherits from that, it can also use it (albeit
having to define its own set_cart_item_order_item_id() method because
it stores the order item ID in different places to renewal handling).

Fixes #2026