Prevent saving subscriptions with a 0 start date
In WC 3.0 a lot of WC core functions expect a datetime object to be
returned from get_date_created(). If a subscriptions start date is
set to 0 this will cause null to be returned and lead to fatal errors.

To prevent that from happening admin shouldn't be able to enter a start date
of 0.