Don't clear upgrade log
Keep a record indefinitely of which versions were updated when. It was
only the 2.0.0 upgrade log that is worth clearing because it logged so
many things. For all other upgrades, its handy to know when the store
updated. If a store is updating to 2.0 from now on, it will also be very
handy to keep that log around.