Use the date created or date paid when calculating next payment date
Which ever date occurred later. Some subscription orders are paid
some time after they are created (manual renewals or failed parent orders)
this means we may need to use the paid date to calculate the next payment
date.

Prior to 2.2.0 the renewal order's post_dates were updated to the current
time which corresponds with values returned by the last_order_date_created
date key. However in 2.2 when a renewal is completed we only set the paid date.

This patch ensures we use the paid date where it occurs after the created date.