Fix inconsistency in var type returned by get_billing_interval
In WC 2.6.14 the value returned is pulled directly from post meta, however,
in WC 3.0+ the value returned is the value stored in memory set by the prop
setter. This creates an inconsistency because the value in the DB is a string
but the value stored in memory is an int.