Store coupon properties in session rather than full coupon object
Storing the full coupon object serialises the data, meaning when it is
loaded and unserialize WC_Data::__wakeup() is called which calls
WC_Coupon::__construct. Because we unserialize the object inside the
construct function it causes an infinite loop. To fix this we only
need to store the coupon properties so we can reconstruct the coupon
from the data.

See https://github.com/Prospress/woocommerce-subscriptions/issues/2106#issuecomment-299098991
for further explanation