Repair pending cancelled subscriptions as part of the 2.2.7 upgrade
WCS 2.2.0 introduced an issue where cancelling a subscription wouldn't
schedule the end of prepaid term scheduled action. This leads to subscriptions
not automatically cancelling automatically on its end date. To fix this we either
need to schedule a new scheduled action or cancel the subscription, depending on whether
the end date is still in the future.