Instantiate a new subscription before scheduling dates on status update
When a subscription's status is updated a bunch of data is also  updated,
for example deleting retry dates etc. Sometimes this happens on a different
instance of the subscription which means functions hooked in later become
out of sync and are using an object which is now out of date.

This commit attempts to fix that by getting a new instance of the object at
the start of WCS_Action_Scheduler::update_status before dates are rescheduled.