Register subscription report scripts on toplevel_page_wc screens
User's which have the view_woocommerce_reports capabilities can access
a limited admin dashboard with WooCommerce sales reports. On this view,
the screen id is different to normal admin users and so we need to also
load our report scripts on that screen as well. This is similar to how
WC core does it.