Fix request param for DoReferenceTransaction
The DoReferenceTransaction API request needs to use deprecated
parameters, meaning L_PAYMENTREQUEST_0_AMT0 isn't used as the key
for its total. The AMT key is used instead. That means when
manually adjusting the total for rounding errors, that key needs
to be used instead of L_PAYMENTREQUEST_0_AMT0.

Using the wrong key leads to:
* PHP Undefined index notice; and
* PayPal Warning in response: "The totals of the cart item amounts
  do not match order amounts", becuase we're not actually fixing
  the rounding error

Related to #2213