Fix wcs_get_product_limitation() default
When SHA: 117aa04c9f1f updated wcs_get_product_limitation() to use CRUD,
it changed it to use an incorrect default return value - 0 instead of
'no'. This fixes that.

This fixes the issue reported in #2218. There will likely be other
implications also as wcs_get_product_limitation() is used in a large
number of places and compares the return value to a variety of things
(e.g. every comparison of != 'no' would have been returning true when
the value wasn't set, when it should have been returning false.

Fixes #2218 and probably a few other things.