Remove old PayPal IPN payload handling
The WCS_PayPal_Standard_IPN_Handler::get_order_id_and_key() method
had code supporting IPN payloads as far back as WC 1.5. However,
since #797 reintroduced the use of matching the PayPal subscription
ID via post meta (which was first introduced with SHA: ba63a40), this
should almost always be unnecessary once the first IPN is processed
and the PayPal subscription ID is set on the subscription (and we only
need to handle that first IPN with supported versions of WC, which is
v2.5+ at this time, all which consistently use JSON data and which is
explicitly set by Subscriptions now in WCS_PayPal_Standard_Request).

Because of that, it's safe to remove this legacy code.

In the rare cases where it's not safe it will be because the data has
been corrupted on the subscription and '_paypal_subscription_id' is
not set correctly. In those cases, we can now throw an exception and
let the  WCS_PayPal_Standard_IPN_Failure_Handler class handle it and
notify the store manager to contact support.