Add new "Create Parent Order" action
For creating an initial order for the customer to pay to activate
a subscription.

To do so, we also need a new 'parent' order type for creation via
wcs_create_order_from_subscription(). We also need to make sure
*no* subscription related meta data is copied to the parent order,
like _requires_manual_renewal, because that will override the meta
data on the subscription created at checkout by
WC_Subscriptions_Checkout::create_subscription(), which calls
wcs_copy_order_address().

Part of #2123