Format PayPal totals as strings when comparing calculated & order totals
PHP can't compare floats in a reliable way. This leads to line items
being skipped (sent as 1 line item) when it's not necessary. Floats
despite being rounded still have a large number of decimals which can
cause comparisons to return false where they are effectively the same
price.

See
http://php.net/manual/en/language.types.float.php#language.types.float.comparison

To get around that, this PR introduces `price_format()` based of
WooCommerce's `number_format()` which returns the prices as strings
which can be used for comparisons reliably.