Fix cancellation IPN handling after payment change
When a renewal payment fails with PayPal Standard, if the customer pays
for the failed renewal order using PayPal Standard again, Subscriptions
will cancel the old subscription, which triggers both 'subscr_cancel'
and 'subscr_eot' IPNs. However, because Subscriptions will also have
updated the '_paypal_subscriber_id' value on the subscription, it won't
find a matching subscription with get_order_id_and_key(), which would
now trigger an exception. However, we don't need an exception for this,
we just want to ignore it, as nothing needs to be done - the subscription
should not be cancelled in WooCommerce, as it has a new, valid PayPal
Standard profile.

Fixes #2283