Set switch order item legacy values property in WC compatible format
WC core sets the line item legacy values property to be the full cart
item
(https://github.com/woocommerce/woocommerce/blob/master/includes/class-wc-checkout.php#L361).
In the case of a pending switch order item we were setting it to be the
product object. This commit means we are compatible with WC core.