Delay granting download permissions until after the switch is complete
Download permissions are granted by WC earlier in the order status
transition process compared to processing the switch request and
updating the subscription line items etc. This means when
WCS_Download_Handler::save_downloadable_product_permissions() trys to
move the downloads from the order to the subscription it gets a
subscription which still contains the old line items. This means the
download permissions aren't granted to the subscription and aren't
revoked from the switch order.

This commit fixes that by ensuring we only attempt to move the
permissions from the order to the subscription after the switch is
complete.