[WC3.2] Add support for applying subscription coupon types to subscriptions and orders (#2371)
* WC3.2 support getting discount amount from line items

WC 3.2 added a feature which allows store admin to apply coupons to
subscriptions and orders from the admin edit screen. To support this we
need to be able to get the discount amount for a line item without using
function which rely on the cart (eg. wcs_cart_contains_renewal()).

This commit refactors WC_Subscriptions_Coupon::get_discount_amount()
into 2 functions for getting the discount amount for line items and cart
items.

* WC3.2 validate coupons being applied to orders/subscriptions by admin

* Display more detailed descriptions for invalid coupons

By default, WC will display "Invalid coupon" when a coupon is found to
be invalid. This commit adds more detailed explanations when a
subscription coupon is found to be invalid.