[WC3.2] Ensure download links are included in order emails
WC 3.2 moved the downloads out of the order items table to a separate
table. This breaks the did action hack we used to determine if download
links should be included. This is because the `is_download_permitted()`
check no longer falls within the two actions being compared. This commit
changes the actions used to use higher level actions which wrap the
entire email template.