Trigger a do_action when we fail to create a renewal order
When an exception is thrown while a renewal order is being created,
it's difficult to get details about what the error is. This is
because none of the normal renewal order filters are called and a
wp_error object is simply returned.

This commit adds a do_action so custom logging plugins we write for
support can get more information without needing to edit the plugin code
directly.