Update the method to check if a payment is up front
When running through this method during development using Xdebug, I noticed that it is called multiple times for the same product in a given request. To save a bit of processing, I set up a static variable to serve as a cache during a single request, so we don't have to process the same thing multiple times.