Only redirect to single sub page on account pages
To fix a perf issue with checkout, caused by subscription_thank_you()
calling wc_get_endpoint_url() to link to the subscriptions account
page, at a time when related orders cache isn't set, and so would
need to be generated (during checkout).

I've gone with restricting this to is_account_page() instead of excluding
it on is_order_received_page() becuase there are plenty of other occassions
where it could lead to perf issues. And really, the purpose of redirecting
directly to the single subscription page is intended for streamlining
the experience on My Account anyway, not overloading *all* calls to
that endpoint URL.