Call Abstract_WC_Order_Data_Store_CPT::update()
Not parent::update(), because WC_Order_Data_Store_CPT includes a JIT
setter of the paid date to update orders created with WC < 3.0 (where
paid date wasn't always set). While this is valuable and important for
orders, it is not needed for subscriptions and also results in resource
intensive operations being undertaken to find that date, which is actually
the most recently paid order's date, and therefore, results in running the
related order query, which is slow, even when there are no renewal orders
we need to check for.