Always save the subscription object created via API request
Subscriptions which are created with an active status which also call
payment complete won't save because there is no status transition
required. Currently, the API is relying on `update_status()` to call
payment complete, however, in this case, that function will exit early.
This leads to payment meta not being saved for example.

This commit makes sure we save everytime irrespective of whether payment
complete will be called. This is similar to what WC core does here:
https://github.com/woocommerce/woocommerce/blob/3.3.3/includes/api/class-wc-rest-orders-controller.php#L522-L529