Order related orders by date and ID
To provide more consistent ordering for multiple orders created
at the same time and also attempt to preserve the order from newest
to oldest, by having higher order IDs first in the array.

This is useful for unit tests, and should have little, if any, perceptable
impact on live sites unless there was a batch order import and it was able
to complete fast enough to add more than one order per second.