Do not pass prev_value when not available
Specifically, do not source the previous value from the current meta
value, because there is no guarantee the meta value is unique and can
be matched based on the first value found.

For example, a renewal order can be linked to multiple subscriptions.
Prior to this patch, the previous relationship would be seen as being
removed/updated when any subsequent relationships are added. What we
want instead is for subsequent relations not to change the any existing
values.