Remove Abstract_ prefix on debug tool classes
Not sure why I added this, it's not convention in either WooCommerce
core or Subscriptions to prefix abstracts like that (although there
are of course some exceptions, like Abstract_WC_Order_Data_Store_CPT
and Abstract_WC_Order_Item_Type_Data_Store).

This can safely be renamed as it's unrelased.