Fix 'Resubscribed Subscription' in meta box
By making sure we call the correct datastore method. SHA: a208a6c40
incorrectly replaced a get_post() call that was querying using the
subscription/post ID as the meta value with the cache call that would
return the results where the subscription/post ID was the meta row's
post_id.

Bonus: now that we have both the resubscribe subscription and order
cached, we can display both on the old subscription with no
performance hit, so this commit also does that.