Don't autocomplete orders which have a subtotal greater than 0
When purchasing a sync'd product, if the order total is $0 it's
autocompleted. This is because we assume there is nothing to ship.
Because we used WC_Order::get_total(), customers who applied a coupon
which brought the order total to $0, were also captured by this. To make
sure we only autocomplete orders containing prorated synced products,
check the order subtotal (combined line item total) instead.