Add WCS_Post_Meta_Cache_Manager_Many_To_One
For handling customer post meta, which we know shoudl only ever have
a single cache (because a subscription can only ever belong to one
user) and potentially other cases where a many-to-one relationship
needs to be enforced for a post meta value.

This approach was originally used in WCS_Post_Meta_Cache_Manager, but
it needed to be removed as renewal orders can be linked to multiple
subscriptions (i.e. a many-to-many relationship). See SHA: 64f2a525cc
for details on its removal.