Introduce get_grouped_parent_product_ids for getting grouped parent IDs
There is a loophole in WC where changing a grouped product to a
different product type won't delete the children meta. If that group was
previously setup with children, the children meta still exists and
causes issues when switching.

In those circumstances, we only want to get parent product IDs which
where the parent still exists, and the parent is a group. This function
will fill that gap.