Fix customer user filter on Subscriptions admin
There appears to be a conflict, possibly only with WC 3.5-beta, with
WC_Admin_List_Table_Orders:query_filters() and the filter it applies
to customer_user, and the equivalent in Subscriptions. Most likely
this is a result of the post_author change, and the databases being
out of sync (as #2559 hasn't been merged yet), but we might as well
use our own filter now that we have a cached source, which may even
be faster than queries on 'post_author'.