Rename $item_data to $product_in_cart
The $item_data variable name was derived from the name of the
data's source - $cart_item['data']. However, 'data' is never a
good name, it's too generic. Item is also polysemous - it could
refer to a cart item, or an order or subscription line item (which
can be fee, product, tax or shipping item type).

Much better to use a variable name describing the contect of the
variation, with is a WC_Product, as derived from the data stored
in the cart.