Validate the update all subscriptions request inline with JS validation
When processing the update all subscription payment method requests,
make sure we use the same conditions that we use to display or hide the checkbox on the form.

This also prevents corner cases where the customer
checks the "update all"  box, changes the payment method to a non-supporting payment method,
completes the process and the form actually attempts to handle their request to update all.